requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
django-dbbackup==3.3.0
